refactor: moved structures to a specific file, separating structures from projectsclient implementation
